migrations/pyproject.toml
[tool.poetry]
name = "migrations"
version = "0.1.0"
description = "Database migrations for AI Resume Creator"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.12.0"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
pgvector = "^0.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

